fixed_projected_artifacts/src/streamlit/openflow_quickstart_app.py
#!/usr/bin/env python3
"""Generated from model-driven projection"""

import os
import time
import redis
import jwt
import streamlit as st
import plotly.graph_objects as go
from datetime import datetime, timezone, timedelta
from dataclasses import dataclass
from typing import Dict, Optional, List
from pydantic import BaseModel, Field, field_validator
from cryptography.fernet import Fernet
import boto3
from botocore.exceptions import ClientError
import re
import html

SECURITY_CONFIG = {'fernet_key': os.getenv('FERNET_KEY', Fernet.generate_key()), 'redis_url': os.getenv('REDIS_URL', 'redis://localhost:6379'), 'jwt_secret': os.getenv('JWT_SECRET', 'your-secret-key'), 'session_timeout_minutes': int(os.getenv('SESSION_TIMEOUT_MINUTES', '15')), 'max_login_attempts': int(os.getenv('MAX_LOGIN_ATTEMPTS', '3')), 'password_min_length': int(os.getenv('PASSWORD_MIN_LENGTH', '12'))}
AWS_CONFIG = {'region': os.getenv('AWS_REGION', 'us-east-1'), 'access_key': os.getenv('AWS_ACCESS_KEY_ID'), 'secret_key': os.getenv('AWS_SECRET_ACCESS_KEY')}


if __name__ == "__main__":
    main()
